+++ Assets/Scripts/Creature/Genetics/DNA.cs
