+++ Assets/Scripts/Creature/Genetics/Pattern.cs.meta
