+++ Assets/Scripts/Utility/Editor/Utility.Editor.asmdef
@@ -1,0 +1,18 @@
+{
+    "name": "Utility.Editor",
+    "rootNamespace": "Utility.Editor",
+    "references": [
+        "GUID:c5e212f562927a742a428d3247cbac75"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
