+++ ProjectSettings/ProjectSettings.asset
@@ -137,6 +137,7 @@ PlayerSettings:
   bundleVersion: 0.1
   preloadedAssets:
   - {fileID: -8426688158082476039, guid: 3a8406893a5cc3041b9b79fc925a14ce, type: 2}
+  - {fileID: 0}
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
